Reworked the landscape update system so it is more performant which also meant changing the way grass gets moistened

+++ common/src/game/fire.h
@@ -7,8 +7,8 @@ extern uint fireCount;
 void fireNew           (u16 x, u16 y, u16 z, i16 strength);
 void fireDel           (uint i);
 fire *fireGetAtPos     (u16 x,u16 y, u16 z);
-void fireBox           (int x, int y, int z, int w, int h, int d, int strength);
-void fireBoxExtinguish (int x, int y, int z, int w, int h, int d, int strength);
+void fireBox           (u16 x, u16 y, u16 z, u16 w, u16 h, u16 d, int strength);
+void fireBoxExtinguish (u16 x, u16 y, u16 z, u16 w, u16 h, u16 d, int strength);
 void fireSendUpdate    (uint c, uint i);
 
 fire *fireGetByBeing   (being b);
